safari-600.7-branch/LayoutTests/fast/dom/features-expected.txt

"Core", undefined: true, true
"HTML", undefined: true, true
"XML", undefined: true, true
"Views", undefined: true, true
"StyleSheets", undefined: true, true
"CSS", undefined: true, true
"CSS2", undefined: true, true
"CSS3", undefined: true, true
"Events", undefined: true, true
"UIEvents", undefined: true, true
"MouseEvents", undefined: true, true
"HTMLEvents", undefined: true, true
"MutationEvents", undefined: true, true
"Range", undefined: true, true
"Traversal", undefined: true, true
"Core", null: true, true
"HTML", null: true, true
"XML", null: true, true
"Views", null: true, true
"StyleSheets", null: true, true
"CSS", null: true, true
"CSS2", null: true, true
"CSS3", null: true, true
"Events", null: true, true
"UIEvents", null: true, true
"MouseEvents", null: true, true
"HTMLEvents", null: true, true
"MutationEvents", null: true, true
"Range", null: true, true
"Traversal", null: true, true
"Core", "": true, true
"HTML", "": true, true
"XML", "": true, true
"Views", "": true, true
"StyleSheets", "": true, true
"CSS", "": true, true
"CSS2", "": true, true
"CSS3", "": true, true
"Events", "": true, true
"UIEvents", "": true, true
"MouseEvents", "": true, true
"HTMLEvents", "": true, true
"MutationEvents", "": true, true
"Range", "": true, true
"Traversal", "": true, true
"Core", "1.0": true, true
"HTML", "1.0": true, true
"XML", "1.0": true, true
"Views", "1.0": true, true
"StyleSheets", "1.0": true, true
"CSS", "1.0": true, true
"CSS2", "1.0": true, true
"CSS3", "1.0": true, true
"Events", "1.0": true, true
"UIEvents", "1.0": true, true
"MouseEvents", "1.0": true, true
"HTMLEvents", "1.0": true, true
"MutationEvents", "1.0": true, true
"Range", "1.0": true, true
"Traversal", "1.0": true, true
"Core", "2.0": true, true
"HTML", "2.0": true, true
"XML", "2.0": true, true
"Views", "2.0": true, true
"StyleSheets", "2.0": true, true
"CSS", "2.0": true, true
"CSS2", "2.0": true, true
"CSS3", "2.0": true, true
"Events", "2.0": true, true
"UIEvents", "2.0": true, true
"MouseEvents", "2.0": true, true
"HTMLEvents", "2.0": true, true
"MutationEvents", "2.0": true, true
"Range", "2.0": true, true
"Traversal", "2.0": true, true
"Core", "3.0": true, true
"HTML", "3.0": true, true
"XML", "3.0": true, true
"Views", "3.0": true, true
"StyleSheets", "3.0": true, true
"CSS", "3.0": true, true
"CSS2", "3.0": true, true
"CSS3", "3.0": true, true
"Events", "3.0": true, true
"UIEvents", "3.0": true, true
"MouseEvents", "3.0": true, true
"HTMLEvents", "3.0": true, true
"MutationEvents", "3.0": true, true
"Range", "3.0": true, true
"Traversal", "3.0": true, true
"Core", "nonsense": true, true
"HTML", "nonsense": true, true
"XML", "nonsense": true, true
"Views", "nonsense": true, true
"StyleSheets", "nonsense": true, true
"CSS", "nonsense": true, true
"CSS2", "nonsense": true, true
"CSS3", "nonsense": true, true
"Events", "nonsense": true, true
"UIEvents", "nonsense": true, true
"MouseEvents", "nonsense": true, true
"HTMLEvents", "nonsense": true, true
"MutationEvents", "nonsense": true, true
"Range", "nonsense": true, true
"Traversal", "nonsense": true, true
undefined, "": true, true
undefined, "1.0": true, true
undefined, "2.0": true, true
undefined, "3.0": true, true
undefined, "nonsense": true, true
undefined, undefined: true, true
undefined, null: true, true
null, "": true, true
null, "1.0": true, true
null, "2.0": true, true
null, "3.0": true, true
null, "nonsense": true, true
null, undefined: true, true
null, null: true, true
"nonsense", "": true, true
"nonsense", "1.0": true, true
"nonsense", "2.0": true, true
"nonsense", "3.0": true, true
"nonsense", "nonsense": true, true
"nonsense", undefined: true, true
"nonsense", null: true, true
